feat(operator): changed return type of start_operator. started complete rework
WARNING: This commit removes code and sets back progress.

- Changed return type from an int to a structure, to allow the Operator PID to be returned to the
main server process, for the shutdown procedure.
- Started a complete rework of the Operator process code to remove the usage of polling when
reading the requests from both the main server process and the worker processes.

BREAKING CHANGE: The Operator will be subject to a complete rework due to a core misuse of functions
not supposed to be used on this project.